backend/requirements.txt
asgiref==3.2.10
Django==2.2.16
django-filter==21.1
djangorestframework==3.12.4
gunicorn==20.1.0
psycopg2-binary==2.8.6
PyJWT==2.1.0
pytz==2020.1
sqlparse==0.3.1
python-dotenv==0.20.0
djoser==2.1.0
drf-base64==2.0
drf-extra-fields==3.2.1
Pillow==9.3.0

